pyproject.toml
[tool.poetry]
name = "ztrack"
version = "0.1.0"
description = "A cross-platform time-tracking tool that sits in the system tray."
authors = ["mmatlin <33261504+mmatlin@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pystray = "^0.17.1"
appdirs = "^1.4.4"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[tool.poetry.plugins."gui_scripts"]
ztrack = "ztrack.__main__:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
